requirements.txt
bandit==1.7.4
black==22.10.0
click==8.1.3
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.29
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
pbr==5.11.0
platformdirs==2.5.3
python-dotenv==0.21.0
PyYAML==6.0
smmap==5.0.0
stevedore==4.1.0
tomli==2.0.1
watchdog==2.1.9
Werkzeug==2.2.2
